Styled Card to look like General Widget
